+++ tasks/Gemfile
@@ -26,6 +26,12 @@ gem 'jbuilder', '~> 2.7'
 # Reduces boot times through caching; required in config/boot.rb
 gem 'bootsnap', '>= 1.4.4', require: false
 
+# Kafka
+gem 'delivery_boy'
+gem 'racecar', github: 'zendesk/racecar'
+
+gem 'oj'
+
 #Events
 gem 'wisper'
 gem 'wisper_kafka', github: 'bookmate/wisper_kafka'
