+++ tasks/config/delivery_boy.yml
@@ -1,15 +1,13 @@
 # These config values will be shared by all environments but can be overridden.
-defaults: &defaults
-  client_id: 'uberpopug'
-  compression_codec: 'gzip'
-  # compression_threshold: 1 # default value
+common: &common
+  client_id: "tasks"
 
 development:
-  <<: *defaults
+  <<: *common
   brokers:
-    - localhost:29092
+    - localhost:9092
 
 test:
-  <<: *defaults
+  <<: *common
   brokers:
-    - localhost:29092
+    - localhost:9092
