--- tasks/config/initializers/kafka.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-raise 'Statsd is not initialized' unless defined?(STATSD)
-
-require 'kafka/datadog'
-Kafka::Datadog.statsd = STATSD
