+++ tasks/config/initializers/wisper.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-WisperKafka::Settings.topic = 'wisper_events'
-
 Wisper.subscribe(AccountListener, 
   on: [
     :account_updated,
