+++ tasks/config/racecar.yml
@@ -1,19 +1,12 @@
-# These config values will be shared by all environments but can be overridden.
 common: &common
-  client_id: "bookmate"
-  session_timeout: 30 # 30 seconds
-  heartbeat_interval: 10
-  max_poll_interval: 1200 # 20 minutes
-  pause_timeout: 10 # for 1.x compatibility, pause 10s when got an exception and retry forever
+  client_id: "tasks"
 
 development:
   <<: *common
   brokers:
-    - 'localhost:29092'
-  log_level: debug
+    - localhost:29092
 
 test:
   <<: *common
   brokers:
     - localhost:29092
-  log_level: debug
